тестирование с redux
