+++ package.json
@@ -36,6 +36,7 @@
     "react-dom": "16.13.1",
     "react-redux": "7.2.1",
     "react-router-dom": "5.2.0",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "redux-mock-store": "^1.5.4"
   }
 }
